+++ alacritty/src/config/mod.rs
@@ -100,17 +100,21 @@ pub fn load(options: &Options) -> Config {
     let config_options = options.config_options().clone();
     let config_path = options.config_path().or_else(installed_config);
 
-    if config_path.is_none() {
-        info!(target: LOG_TARGET_CONFIG, "No config file found; using default");
-    }
-
     // Load the config using the following fallback behavior:
     //  - Config path + CLI overrides
     //  - CLI overrides
     //  - Default
     let mut config = config_path
-        .and_then(|config_path| load_from(&config_path, config_options.clone()).ok())
-        .unwrap_or_else(|| Config::deserialize(config_options).unwrap_or_default());
+        .as_ref()
+        .and_then(|config_path| load_from(config_path, config_options.clone()).ok())
+        .unwrap_or_else(|| {
+            let mut config = Config::deserialize(config_options).unwrap_or_default();
+            match config_path {
+                Some(config_path) => config.ui_config.config_paths.push(config_path),
+                None => info!(target: LOG_TARGET_CONFIG, "No config file found; using default"),
+            }
+            config
+        });
 
     // Override config with CLI options.
     options.override_config(&mut config);